Refactor waitCRDs function
Move the logic to wait for CRDs to pkg/manager package to keep main.go
slightly cleaner.
