+++ cmd/tetragon/main.go
@@ -66,12 +66,8 @@ import (
 	"github.com/spf13/viper"
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/durationpb"
-	v1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	apiextensionsclientset "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
-	apiextensionsinformer "k8s.io/apiextensions-apiserver/pkg/client/informers/externalversions/apiextensions/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/cache"
 )
 
 var (
@@ -401,8 +397,22 @@ func tetragonExecuteCtx(ctx context.Context, cancel context.CancelFunc, ready fu
 		// Start controller-runtime manager.
 		controllerManager = manager.Get()
 		controllerManager.Start(ctx)
+		crds := make(map[string]struct{})
+		if option.Config.EnableTracingPolicyCRD {
+			crds[v1alpha1.TPName] = struct{}{}
+			crds[v1alpha1.TPNamespacedName] = struct{}{}
+		}
+		if option.Config.EnablePodInfo {
+			crds[v1alpha1.PIName] = struct{}{}
+		}
+		if len(crds) > 0 {
+			err = controllerManager.WaitCRDs(ctx, crds)
+			if err != nil {
+				return err
+			}
+		}
 		// retrieve k8s clients
-		k8sClient, _, err = watcher.GetK8sClients(waitCRDs)
+		k8sClient, _, err = watcher.GetK8sClients(func(_ *rest.Config) error { return nil })
 		if err != nil {
 			return err
 		}
@@ -537,57 +547,6 @@ func tetragonExecuteCtx(ctx context.Context, cancel context.CancelFunc, ready fu
 	}
 
 	return obs.StartReady(ctx, ready)
-}
-
-func waitCRDs(config *rest.Config) error {
-	crds := make(map[string]struct{})
-
-	if option.Config.EnableTracingPolicyCRD {
-		crds[v1alpha1.TPName] = struct{}{}
-		crds[v1alpha1.TPNamespacedName] = struct{}{}
-	}
-	if option.Config.EnablePodInfo {
-		crds[v1alpha1.PIName] = struct{}{}
-	}
-
-	if len(crds) == 0 {
-		log.Info("No CRDs are enabled")
-		return nil
-	}
-
-	log.WithField("crds", crds).Info("Waiting for required CRDs")
-	var wg sync.WaitGroup
-	wg.Add(1)
-	crdClient := apiextensionsclientset.NewForConfigOrDie(config)
-	crdInformer := apiextensionsinformer.NewCustomResourceDefinitionInformer(crdClient, 0*time.Second, nil)
-	_, err := crdInformer.AddEventHandler(cache.ResourceEventHandlerFuncs{
-		AddFunc: func(obj interface{}) {
-			crdObject, ok := obj.(*v1.CustomResourceDefinition)
-			if !ok {
-				log.WithField("obj", obj).Warn("Received an invalid object")
-				return
-			}
-			if _, ok := crds[crdObject.Name]; ok {
-				log.WithField("crd", crdObject.Name).Info("Found CRD")
-				delete(crds, crdObject.Name)
-				if len(crds) == 0 {
-					log.Info("Found all the required CRDs")
-					wg.Done()
-				}
-			}
-		},
-	})
-	if err != nil {
-		log.WithError(err).Error("failed to add event handler")
-		return err
-	}
-	stop := make(chan struct{})
-	go func() {
-		crdInformer.Run(stop)
-	}()
-	wg.Wait()
-	close(stop)
-	return nil
 }
 
 func loadTpFromDir(ctx context.Context, dir string) error {
